\docType{package}
\name{uaparserjs}
\alias{uaparserjs-package}
\alias{uaparserjs}
\title{Parse 'User-Agent' Strings}
\description{
This package parses HTTP user agent strings and returns user agent, device and OS information.  
It does this by providing a wrapper around the user agent parser provided by 
the ua-parser project's Javascript reference parser (<https://github.com/ua-parser>)
}
\details{
The package incorporates regular expressions and test data from the ua parser project 
that are copyrighted by Google and the UA Parser contributors and are used 
under an Apache 2 license.  See the UA parser project for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gregfrog/uaparserjs}
  \item Report bugs at \url{https://github.com/gregfrog/uaparserjs/issues}
}

}
\author{
Bob Rudis (@hrbrmstr)
Greg Hunt
}
