% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uaparserjs-package.R
\docType{package}
\name{uaparserjs}
\alias{uaparserjs}
\alias{uaparserjs-package}
\title{Parse Browser User Agent Strings into Data Frames}
\description{
Parse Browser User Agent Strings into Data Frames
}
\author{
Bob Rudis (@hrbrmstr)
}

