% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uaparserjs-package.R
\docType{package}
\name{uaparserjs}
\alias{uaparserjs}
\title{Parse 'User-Agent' Strings}
\description{
Despite there being a section in RFC 7231
<https://tools.ietf.org/html/rfc7231#section-5.5.3> defining a suggested
structure for 'User-Agent' headers this data is notoriously difficult
to parse consistently. Tools are provided that will take in user agent
strings and return structured R objects. This is a 'V8'-backed package
based on the 'ua-parser' project <https://github.com/ua-parser>.
}
\author{
Bob Rudis (@hrbrmstr)
}
